requirements.txt
environs==11.0.0
python-telegram-bot==13.15
redis==5.0.8
vk-api==11.9.9
python-dotenv~=1.0.1